Added support for File task supplier and made the pipeline non blocking
