+++ service/src/main/java/com/booking/validator/service/supplier/data/source/QueryConnectorsForTask.java
@@ -5,8 +5,6 @@ import com.booking.validator.data.Data;
 import com.booking.validator.data.source.DataSource;
 import com.booking.validator.task.Task;
 import com.booking.validator.task.TaskComparisonResult;
-import com.booking.validator.task.TaskComparisonResultV1;
-import com.booking.validator.task.TaskV1;
 
 import java.util.concurrent.CompletableFuture;
 import java.util.function.Supplier;
@@ -27,9 +25,9 @@ public class QueryConnectorsForTask implements Supplier<CompletableFuture<TaskCo
 
     public CompletableFuture<TaskComparisonResult> get(){
         if (task == null) return CompletableFuture.completedFuture(null);
-        return CompletableFuture.supplyAsync(QueryConnectorsForTask.getSupplier(((TaskV1) task).getSource()))
+        return CompletableFuture.supplyAsync(QueryConnectorsForTask.getSupplier(task.getSource()))
                 .thenCombineAsync(
-                        CompletableFuture.supplyAsync(QueryConnectorsForTask.getSupplier(((TaskV1) task).getTarget())), task::validate)
-                .exceptionally( t -> new TaskComparisonResultV1((TaskV1) task, null, t));
+                        CompletableFuture.supplyAsync(QueryConnectorsForTask.getSupplier(task.getTarget())), task::validate)
+                .exceptionally( t -> new TaskComparisonResult(task, null, t));
     }
 }
