+++ service/src/test/java/CreateTaskV1.java
@@ -6,14 +6,11 @@ import com.booking.validator.data.transformation.TransformationTypes;
 import com.booking.validator.service.supplier.data.source.QueryConnectorsForTask;
 import com.booking.validator.task.Task;
 import com.booking.validator.task.TaskComparisonResult;
-import com.booking.validator.task.TaskComparisonResultV1;
-import com.booking.validator.task.TaskV1;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.Test;
 
 import static org.junit.Assert.*;
 
-import java.io.Console;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -24,14 +21,14 @@ public class CreateTaskV1 {
         data.put("a", "a");
         DataSource dataSource = new DataSource(
                 "constantSource",
-                new ConstantQueryOptions(Types.CONSTANT.getValue(), (Map<String, Object>)data, null));
-        Task taskV1 = new TaskV1("unit_test", dataSource, dataSource, null);
+                new ConstantQueryOptions(Types.CONSTANT.getValue(), data, null));
+        Task taskV1 = new Task("unit_test", dataSource, dataSource, null);
         System.out.println(taskV1.toJson());
         ActiveDataSourceConnections.getInstance().add("constantSource", Types.CONSTANT.getValue(), null);
         TaskComparisonResult r1 = taskV1.validate(ActiveDataSourceConnections.getInstance().query(dataSource),
                                                   ActiveDataSourceConnections.getInstance().query(dataSource));
         ObjectMapper mapper = new ObjectMapper();
-        TaskV1 pls = mapper.readValue(taskV1.toJson(), TaskV1.class);
+        Task pls = mapper.readValue(taskV1.toJson(), Task.class);
         assertEquals(((ConstantQueryOptions)pls.getSource().getOptions()).getData().get("a"), "a");
     }
 
@@ -50,11 +47,11 @@ public class CreateTaskV1 {
         DataSource dataSource2 = new DataSource(
                 "constantSource",
                 new ConstantQueryOptions("constant", (Map<String, Object>)data2, null));
-        Task taskV1 = new TaskV1("unit_test", dataSource, dataSource2, null);
+        Task taskV1 = new Task("unit_test", dataSource, dataSource2, null);
         System.out.println(taskV1.toJson());
         ActiveDataSourceConnections.getInstance().add("constantSource", Types.CONSTANT.getValue(), null);
         QueryConnectorsForTask qcft = new QueryConnectorsForTask(taskV1);
-        TaskComparisonResultV1 taskComparisonResult = (TaskComparisonResultV1) qcft.get().get();
+        TaskComparisonResult taskComparisonResult = qcft.get().get();
         assertTrue(taskComparisonResult.isOk());
     }
