+++ service/src/test/java/FunctionalTest.java
@@ -8,8 +8,6 @@ import com.booking.validator.service.supplier.task.cli.CommandLineTaskSupplier;
 import com.booking.validator.service.supplier.task.kafka.KafkaTaskSupplier;
 import com.booking.validator.task.Task;
 import com.booking.validator.task.TaskComparisonResult;
-import com.booking.validator.task.TaskComparisonResultV1;
-import com.booking.validator.task.TaskV1;
 import com.booking.validator.utils.Service;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -112,8 +110,8 @@ public class FunctionalTest {
         DataSource dataSource = new DataSource(
                 "constantSource",
                 new ConstantQueryOptions(Types.CONSTANT.getValue(), (Map<String, Object>)data, null));
-        Task taskV1 = new TaskV1("unit_test", dataSource, dataSource, null);
-        return mapper.writeValueAsString(taskV1);
+        Task task = new Task("unit_test", dataSource, dataSource, null);
+        return mapper.writeValueAsString(task);
     }
 
     private Service getKafkaProducer() {
@@ -183,7 +181,7 @@ public class FunctionalTest {
             if (t!=null) {
                 System.out.println(t);
             } else {
-                System.out.println(((TaskComparisonResultV1)x).isOk());
+                System.out.println(x.isOk());
             }
         };
